requirements.txt
numpy
scipy
sklearn
joblib
tqdm
click
opendr
chumpy
pillow
cython
fasteners
h5py
pymp-pypi
plyfile
git+https://github.com/classner/clustertools.git#egg=clustertools
